requirements.txt
# DataPace — Dependencies
# pip install -r requirements.txt

# Core
pandas>=2.0
openpyxl>=3.1
python-dotenv>=1.0

# Validation
pydantic>=2.0

# Testing
pytest>=7.0
pytest-cov>=4.0

# Crawlers (optional)
# playwright>=1.40  # for scraping, install separately: pip install playwright && playwright install
